backend/requirements.txt
flask
flask-socketio
flask-cors
paho-mqtt
opencv-python
cv2-enumerate-cameras
pyzbar
numpy
qreader
requests